requirements-dev.txt
Flask==2.0.3
pytest==7.0.1
coverage==6.3.2
Faker==13.2.0
requests==2.27.1
instagrapi==1.16.14
Pillow==9.0.1